gallium: add renderonly driver
This commit adds a generic renderonly driver library, which fullfille
the requirements for tegra and etnaviv. As a result it is possible to
run unmodified egl software directly (without any compositor) on
supported devices.

In every use case we import a dumb buffer from scanout gpu into
the renderonly gpu.

If the scanout hardware does support the used tiling format from the
renderonly gpu, a driver can define a function which is used to 'setup'
the needed tiling on that imported buffer. This functions gets called
during rendertarget resource creation.

If the scanout hardware does not support the used tiling format we need
to create an extra rendertarget resource for the renderonly gpu.
During XXX we blit the renderonly rendertarget onto the imported dumb
buffer.

We assume that the renderonly driver provides a blit function that is
capable of resolving the tilied into untiled one.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>