etnaviv: caps: add vertex_max_elements capability and use it
Mesa can ask us to create up to PIPE_MAX_ELEMENTS vertex elements,
which always exceeds the available register space.  The available
register space depends on the HALTI0 feature bit, which, when set,
indicates that we can support up to 16 vertex element configurations.

Etna_viv documentation is confused about the meaning when HALTI0 is
clear; in one place it says 12 and another it says 10.  Opt for the
safe value of 10 until this is clarified.

Refuse to create a vertex element configuration which violates this
capability.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>