etnaviv: abort() on unsupported index sizes
GC600 does not support 32-bit index sizes.

If we get an index size we don't support, rather than blindly trying to
emit operations to use it, or emitting an invalid configuration, report
the condition and abort() rather than potentially hanging the GPU.  The
code structure does not allow us to error out at this point.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>