etnaviv: fix GPU hang: support 1D textures
Currently, unsupported texture target modes cause us to issue a texture
type of 7, which causes the GPU to hang.  We need to avoid this.

This commonly occurs when the application tries to use a 1D texture.
Vivante GPUs do not support 1D textures directly, but as 1D textures
always have a height of 1, they can be supported by specifying a 2D
texture and forcing the T coordinate (or V in Vivante speak) to wrap.

Implement this, and ensure that an attempt to use an unsupported
texture type fails (by returning NULL from etna_create_sampler_view()).

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>