etnaviv: add padding for resources created from handles
Initialise the padded_width and padded_height members of resources
obtained from handles.  These must follow the requirements of the
resolve engine, as we will use the RS to copy to/from these resources.
As has been previously established, the RS engine requires a width
alignment of 16 pixels and a height of 4 pixels, or 8 if we have a
dual pixel pipe GPU.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>