etnaviv: fix TGSI_OPCODE_LIT implementation
The TGSI documentation says the LIT maths should be:
	dst.x = 1
	dst.y = max(src.x, 0)
	dst.z = (src.x > 0) ? max(src.y, 0)^{clamp(src.w, -128, 128))} : 0
	dst.w = 1

However, our implementation fails on two points.  Firstly, the handling
of src.y <= 0 is incorrect, and results in dst.z being zero.  Secondly,
src.y is passed to LOG() which produces undefined results for values
less than zero.  Thirdly, we don't clamp src.w to -128..128.

Fix all these issues.  We optimise by checking what kind of source src
is, and if it's a uniform, we apply the limits at compile time.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>