etnaviv: fix write disable to the color buffer
The code in etna_set_framebuffer_state() assumes that clearing the
framebuffer's PE_COLOR_FORMAT state might stop us writing to the color
target.  However, as we combine this with the blend state, we end up
with a state value of 0x00010f00, which has the overwrite and full
color components set, but with a format of X4R4G4B4.  This does not
prevent writing to the color buffer.

Fix this by having these fields in the framebuffer state be a mask for
the blend state: the overwrite and component bits must be all ones for
the blend state to be propagated.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>