etnaviv: purge stale vertex buffers
We currently leave stale vertex buffers BOs and their configuration
behind when vertex buffers are removed. This causes problems with
etnaviv DRM as it causes stale relocations to be passed to the kernel
which can cause the command stream submission to be rejected:

[drm:etnaviv_ioctl_gem_submit] *ERROR* relocation 4 outside object

or worse, we could reference a stale buffer object:

[drm:etnaviv_ioctl_gem_submit] *ERROR* invalid handle 11 at index 3

Avoid this by ensuring that we remove references to stale buffer
objects when the resources go away.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>