etnaviv: only emit states with valid buffer objects
As the kernel command stream parser has been tightened, we need to
avoid writing to states which are "sensitive" without a relocation
present, iow the address registers. The case where we do this is
where we have an address register but no BO to place there.  Rather
than emitting a state, avoid skip emitting it.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>