etnaviv: fix blitting multi and supermultitiled buffers
In the case of a GPU with two pixel pipes using multi-tiled buffers,
the buffer is split into two halves, each half for a pixel pipe.  The
boundary between the two halves is fixed, and does not depend on the
height that we're blitting.

For example, a 1024 line buffer is split into two separate 512 line
buffers.  If we want the top 256 lines, then the existing code sets the
first buffer address to the start, and the second buffer address to 128
lines into the buffer.  This results in fetching lines 0-127 as the
upper half, and 128-255 for the lower half.

However, if we then do a blit for the whole buffer of 1024 lines, the
upper half will be 0-511 and the lower half will be 512-1023.

Obviously, this is incorrect, for the 256 line case, we should fetch
lines 0-127 and 512-639.

Fix this by passing in the padded height of the buffer to
etna_rs_compile_state() so that it can correctly calculate the start
of the second half of the buffer.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>