etnaviv: improve tiling format selection
We have two competing tiling formats: the tiling format for sampling
and the tiling format for render targets.

For older GPUs, these are the same (a plain 4x4 tile) but for more
recent GPUs, the rendering tile format is different from the texture
format.  Moreover, there are formats (eg, byte block sized formats
such as L8/I8 etc) which we are unable to convert using the RS.

Change the tiling format selection to choose more appropriately, now
that we support converting between the rendering and sampling on the
fly.  Our strategy here is:

- If the resource is a buffer, use linear format, otherwise:
- Select the tiling format for rendering if the resource will be used
  for rendering.  (Byte block sized formats can't be used as a
  render target.)
- Select the sampling tiling format if the resource is a byte block
  sized format, and it isn't a depth/stencil buffer.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>