etnaviv: use own fences for transfer map
Use our own fence support for the transfer map operation, rather than
calling etna_cmd_stream_finish().  This allows us to consistently
insert GPU cache flushes prior to submitting the rendering operation
so that we can be sure that the render buffers are properly updated
when the fence has completed.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>