etnaviv: implement transfer_{map,unmap}() tile conversion on GPU
We need to de-tile and de-tile-status a buffer before providing access
to it.  Rather than implementing the multiple different tiling formats
in software, along with the tile status, use the GPUs RS engine to
format a linear buffer.

We avoid this when using the texture buffer, as we can already de-tile
and re-tile the texture buffer.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>