etnaviv: disable tile status
Tile status appears to be broken on GC600 - investigation shows that
the tile status buffer appears to be completely ignored by the GPU
both in render and resolve operations.

This has been proven by switching etnaviv to allocate tile status for
an X8R8G8B8 colour buffer, setting the tile status to 0x11111111, and
the colour buffer to 0xe5e5e5e5.  The clear value is set to 0xff000000.
Render operations appear to write to the colour buffer, denoted by the
top byte (alpha/X channel) becoming zeroed and the colour values being
altered.  However, the tile status remains 0x11111111.  In other words,
the tile status buffer appears to be ignored.

Resolve behaves the same: with the same setup, using the resolve engine
to copy the buffer results in a mere copy of the buffer without any
filling by the clear value.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>