etnaviv: convert line loop primitives
GC600/GC880/GC2000 do not support line loop primitives.  On at least
GC600, this causes a GPU hang with just the PA and FE modules busy.
There is an undocumented "LINE LOOP" feature bit in minor features 2
which is clear on these cores, so use this to decide whether line
loops are supported, and whether we ask for these primitives to be
converted.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>