etnaviv: ensure we do not break draw state across flushes
Under certain circumstances, etnaviv can break the draw state across a
flush to the GPU.  We must not break the draw state across flushes, as
the GPU can be runtime suspended and loose all previous state, or
another application can alter the GPUs state.

Arrange to reserve the full state up front to avoid this.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>