GBM: add KMS provider interface
The GBM device, which is used to bootstrap the EGL context
should always be constructed atop a render capable DRM device.
If this device is a render node or a standalone renderonly DRM
device, we need a way to set the actual KMS output device to
allow for proper allocations of the buffer objects backing the
window surface and hiding the needed PRIME import/export.

This interface allows this by constructing a second GBM device
on top of the KMS capable DRM device and binding this as the
KMS provider to the render GBM device.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>