etnaviv: fix point size - state is radius of points
The point size specified by OpenGL is the diameter of points, but the
GPU's point size state appears to be the radius of the point. Hence,
we need to halve the point diameter before programming the hardware.
This gives the same point size as the same value passed as a point
size from the vertex shader.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Wladimir J. van der Laan <laanwj@gmail.com>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>