etnaviv: add SIN, COS and SCS implementations
Some Vivante GPUs do not have native SIN and COS instructions. For
these cases, we provide an implementation based on the fast sine/
cosine implementation found here:

  http://forum.devmaster.net/t/fast-and-accurate-sine-cosine/9648

We extend it to SCS too; there are parts of the calculation for sine
and cosine calculation which can be done in parallel, which is more
efficient than replicating the code for COS and the code for SIN.

However, SCS currently fails to set the output register Z and W
correctly.

This fixes the SIN/COS piglit glean tests on GC600.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Reviewed-by: Christian Gmeiner <christian.gmeiner@gmail.com>