etnaviv: fix MSAA blit of depth/stencil buffers
If we attempt to blit a depth/stencil buffer, the blit operation hangs
the GPU because we write ~0 to the MSAA configuration register.  This
happens because translate_msaa_format() returns ~0 for these formats.

Trap the invalid format before it reaches the GPU, so we avoid hanging
the hardware, and since we already declare a "compatible" format, use
that instead of the resource format when selecting the MSAA format.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>