etnaviv: fix multiple uniform arguments for LRP
If we have multiple source uniforms for the TGSI LRP instruction, we
can overflow the number of inner temporary registers, since we try to
subsitute the second and third uniforms with temporary registers.  It
emit_inst() can't know that replacing the first uniform with a
register would be a better option, so let's so manually in the LRP
implementation.

Take the principle of implementation from XPD, which already does
something similar.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>