etnaviv: add cpu_prep/cpu_fini to transfers
Add cpu_prep/cpu_fini libdrm-etnaviv calls around the transfer of
resources.  This ensures that any in-progress GPU operations on
the resources will have completed prior to access being granted to
them, which is necessary to avoid rendering errors.

We currently "get away" with this by always submitting and waiting
on a commit, but this can be heavy handed both in terms of work and
power, especially if the GPU was already idle.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>