etnaviv: avoid eliminating MOVs with side effects
A MOV instruction can have side effects such as saturating, negating
and taking the absolute of the source register.  In order to safely
eliminate the MOV, we must first check that it has none of these side
effects, otherwise we are changing the behaviour of the resulting code.

Add checks for these side effects, and avoid eliminating such MOVs.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>