etnaviv: create tiled texture-only resources
When a resource is requested which will only be used for texturing, we
should allocate a suitable resource which is directly compatible with
the texture unit. For existing GPUs, this means a tiled resource.

This avoids unnecessary staging resources if it is known that a
resource is only going to be used as a texture up front.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>