etnaviv: better incompatible texture resource handling
The texture resources which can be accepted by the sampler depend on
the HALIGN resource - it is not just a question of whether it is tiled,
but also whether it is appropriately aligned.  This is important with
non-HALIGN GPUs, where the texture must be aligned to a tile, not to
the RS width/height.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>