etnaviv: Get line rendering to work on GC3000
The GC3000 in i.MX6qp refuses to render lines when WIDE_LINE mode
disabled.

This feature has existed for a long time (gc880, gc1000, ...) and
the Vivante driver always uses it when available, even to draw
"non-wide" lines. So should we.

Set the bit (when the feature is available) and the associated state.

Signed-off-by: Wladimir J. van der Laan <laanwj@gmail.com>