etnaviv: Implement primitive restart
Primitive restart is supported on Vivante hardware with the HALTI0
feature.

To support draw_info-dependent state changes, add a function
etna_update_state_for_draw that updates the index buffer state
according to the draw state.

This is one small step on the way to ES3. As there is no ES2 extension
for primitive restart, to test this you currently need to force the GL
ES version to 3 with `export MESA_GLES_VERSION_OVERRIDE=3.0`.

Signed-off-by: Wladimir J. van der Laan <laanwj@gmail.com>